pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lumenguard-modal"
version = "0.1.0"
description = "IP monitoring with Telegram notifications via Modal"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27,<1.0",
  "modal>=1.0,<2.0",
  "pydantic>=2.8,<3.0",
  "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
